% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFA.R
\name{modes}
\alias{modes}
\title{Provides mode for a distribution vector}
\usage{
modes(dis_vec)
}
\arguments{
\item{dis_vec}{A distribution vector.}
}
\value{
mode_dis_vec A distribution vector.
}
\description{
Provides mode for a distribution vector
}
\examples{
Provides a single value if there is only one mode

modes(c(1,3,2,4,5,3)
modes(c(1,2,3))

}
